+++ app/models/pull_request_user.rb
@@ -1,6 +1,6 @@
 class PullRequestUser < ApplicationRecord
   belongs_to :pull_request
-  belongs_to :user
+  belongs_to :user, class_name: 'Contributor'
 
   validates :role, presence: true
 end