+++ app/models/review.rb
@@ -1,6 +1,6 @@
 class Review < ApplicationRecord
   belongs_to :pull_request
-  belongs_to :author, class_name: 'User'
+  belongs_to :author, class_name: 'Contributor'
 
   # Turn on once values are captured for existing review records
   # validates :author, presence: true